+++ demo/CypressLayout.tsx
@@ -31,6 +31,7 @@ import {
   Textarea,
   ToolTip,
   TimePicker,
+  TimeRange,
   Tour,
   Validation,
   VStepper,
@@ -137,6 +138,9 @@ export const CypressLayout = () => (
         </Route>
         <Route path="/cypress/timepicker">
           <TimePicker />
+        </Route>
+        <Route path="/cypress/timerange">
+          <TimeRange />
         </Route>
         <Route path="/cypress/tour">
           <Tour />