+++ demo/cypress/TimeRange/index.tsx
@@ -1,0 +1,21 @@
+import * as React from 'react';
+
+import * as L from '../../../leda';
+
+export const TimeRange = () => {
+  const [value, setValue] = React.useState<[string, string]>(['13:30', '15:00']);
+
+  return (
+    <L.Div _demoStory>
+      <L.TimeRange
+        name="TimeRange"
+        min={new Date(2019, 9, 5, 12, 30)}
+        max={new Date(2019, 9, 7, 17, 0)}
+        value={value}
+        onChange={(ev) => setValue(ev.component.value)}
+        isRequired={[true, false]}
+        _width50
+      />
+    </L.Div>
+  );
+};