+++ demo/cypress/index.ts
@@ -25,6 +25,7 @@ export { RadioGroup } from './RadioGroup';
 export { Slider } from './Slider';
 export { StatusBar } from './StatusBar';
 export { Textarea } from './Textarea';
+export { TimeRange } from './TimeRange';
 export { ToolTip } from './ToolTip';
 export { Tour } from './Tour';
 export { TimePicker } from './TimePicker';