+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [0.38.0] - 2020-12-31
 
 ### Fixed
+- TimeRange: Fixed behaviour on Tab click
 - DateRange: Fixed time for date values